+++ server-data/resources/[bpt_addons]/bpt_ballasjob/locales/en.lua
@@ -1,0 +1,16 @@
+Locales['en'] = {
+  ['amount_invalid'] = 'invalid amount',
+  ['press_to_open'] = 'press [E] to access the menu',
+  ['billing'] = 'billing',
+  ['billing_sent'] = 'the account has been registered!',
+  ['invoice_amount'] = 'invoice amount',
+  ['no_players_near'] = 'no players nearby',
+  ['boss_actions'] = 'Boss actions',
+  ['blip_ballas'] = 'ballas',
+  ['ballas'] = 'ballas',
+  -- Inventory
+  ['deposit_stock'] = 'deposit stock',
+  ['take_stock'] = 'take stock',
+  ['have_deposited'] = 'have deposited',
+  ['quantity_invalid'] = 'quantity invalid'
+}