+++ server-data/resources/[bpt_addons]/bpt_ballasjob/locales/it.lua
@@ -7,5 +7,10 @@ Locales['it'] = {
   ['no_players_near'] = 'nessun giocatore nelle vicinanze',
   ['boss_actions'] = 'Azioni del boss',
   ['blip_ballas'] = 'ballas',
-  ['ballas'] = 'ballas'
+  ['ballas'] = 'ballas',
+  -- Inventory
+  ['deposit_stock'] = 'Depositare',
+  ['take_stock'] = 'Prendi',
+  ['have_deposited'] = 'Hai depositato',
+  ['quantity_invalid'] = 'Qunatità non valida'
 }