+++ server-data/resources/[ox]/ox_inventory/data/stashes.lua
@@ -133,25 +133,6 @@ return {
 	},
 
 	{
-		coords = vec3(78.738464, -1962.184570, 18.041016),
-		target = {
-			loc = vec3(78.738464, -1962.184570, 18.041016),
-			length = 0.6,
-			width = 1.8,
-			heading = 340,
-			minZ = 43.34,
-			maxZ = 44.74,
-			label = 'Storage'
-		},
-		name = 'ballaslocker',
-		label = 'Storage',
-		owner = false,
-		slots = 70,
-		weight = 70000,
-		groups = {['ballas'] = 0}
-	},
-
-	{
 		coords = vec3(-29.287910, -1103.182373, 26.415405),
 		target = {
 			loc = vec3(-29.287910, -1103.182373, 26.415405),