+++ .github/CHANGELOG.md
@@ -74,3 +74,5 @@ fix lint error: unused argument last; accessing undefined variable Invoke; unuse
 47. [server.cfg]: fix: Couldn't find resource @bitpredator
 48. fix: No such export getSharedObject in resource es_extended @bitpredator
 49. [bpt_bakerjob]: refactor: bpt_bakerjob inventory implementation @bitpredator
+50. [bpt_ballasjob]: refactor: bpt_ballasjob inventory implementation @bitpredator
+51. [bpt_ballasjob]: feat: bpt_ballasjob\locales\en.lua @bitpredator